Restructured project directory and moved communication related classes to presenter/communication directory
