+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         android:theme="@style/Theme.AppCompat.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".GameActivity"
+            android:name=".presenter.communication.GameActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
