+++ app/src/main/java/at/vunfer/openrealms/presenter/communication/SenderThread.java
@@ -1,5 +1,5 @@
 /* Licensed under GNU GPL v3.0 (C) 2023 */
-package at.vunfer.openrealms;
+package at.vunfer.openrealms.presenter.communication;
 
 import java.io.PrintWriter;
 import java.util.Random;
