+++ app/src/main/res/layout/activity_game.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".GameActivity">
+    tools:context=".presenter.communication.GameActivity">
 
     <TextView
         android:id="@+id/exampleView"
